Added a scanners folder and renamed scanners
