+++ app.py
@@ -6,8 +6,8 @@ to ensure accessibility standards are met.
 """
 from flask import Flask, request
 from flask_cors import CORS
-from scanner import score_text_contrast
-from textscanner import score_text_accessibility
+from scanners.color_contrast_scanner import score_text_contrast
+from scanners.text_scanner import score_text_accessibility
 
 app = Flask(__name__)
 cors = CORS(
