+++ scanners/text_scanner.py
